rename profiling.py to asv_profiling.py
