+++ asv/commands/asv_profiling.py
@@ -12,7 +12,7 @@ from ..benchmarks import Benchmarks
 from ..console import log
 from ..environment import ExistingEnvironment, get_environments, is_existing_only
 from ..machine import Machine
-from ..profiling import ProfilerGui
+from ..asv_profiling import ProfilerGui
 from ..repo import NoSuchNameError, get_repo
 from ..results import iter_results_for_machine
 from ..runner import run_benchmarks
